build: passer au niveau -Wextra et corriger les warnings associés (#21)
Ajoute -Wextra aux CFLAGS (portable gcc/clang). Corrige les 20 warnings
révélés, dont 2 GCC-only (type-limits) invisibles sous clang :

- unused-parameter (9) : callbacks threads/signaux/stubs marqués `(void)x;`
  (console, etii_client, etii_server, main, possibility).
- sign-compare (8) : conversions rendues explicites (cast) ou compteur de
  boucle aligné en size_t (datamanager, etii_search, lifo, local_socket,
  part, possibility, readdata) — comportement inchangé.
- type-limits (2) : tests « < 0 » sur des uint8_t (possibility.c) supprimés
  car toujours faux.
- stringop-truncation (1) : strncpy -> memcpy dans logger_ncurses.c, comme
  déjà fait dans logger.c.

Vérifié : gcc-15 -Wall -Wextra sans warning (ncurses inclus), build clang,
build WERROR=1 (= ce que lance la CI) et `make test` (134/134) OK.

+++ lifo.c
@@ -265,7 +265,7 @@ int scroll_fifo (File * suite, void *dest){
 	
 	if(inside_cache(suite, supp_element))
 	{
-		if (position_cache(suite, supp_element) == suite->lastPostionCache -1)
+		if ((unsigned long long)position_cache(suite, supp_element) == suite->lastPostionCache -1)
 		{
 			suite->lastPostionCache--;
 		}
